Move changeset history calculation to its own file
